fix: address PR review feedback for delegate creation
- Fix clippy wildcard_enum_match_arm for StreamChunk/StreamHeader
- Clean up DELEGATE_INHERITED_ATTESTATIONS on UnregisterDelegate
- Remove write-only DELEGATE_CREATORS DashMap (dead code)
- Use UnsafeCell for interior mutability in DelegateCallEnv stores
- Make params_len i64 for API consistency (stdlib 0.2.2)
- Add 5 unit tests for resource limit enforcement (depth, per-call,
  invalid WASM, counter tracking)
- Regenerate test-delegate-creation lockfile for stdlib 0.2.2

+++ crates/core/src/wasm_runtime/engine/wasmtime_engine.rs
@@ -1104,7 +1104,7 @@ impl WasmtimeEngine {
                     nonce_ptr,
                     out_key_ptr,
                     out_hash_ptr,
-                ): (i64, i64, i64, i32, i64, i64, i64, i64)| {
+                ): (i64, i64, i64, i64, i64, i64, i64, i64)| {
                     let id = native_api::CURRENT_DELEGATE_INSTANCE.with(|c| c.get());
                     refresh_mem_addr_from_caller(&mut caller, id);
                     Box::new(async move {
